+++ src/PublicApi/SubscriptionEndpoints/ListUserSubscriptionsEndpoint.cs
@@ -74,7 +74,7 @@ public partial class ListUserSubscriptionsEndpoint : IEndpoint<IResult, EmptyReq
                     var subscription = subscriptionResponse.Subscription;
                     response.Subscriptions.Add(new SubscriptionDto
                     {
-                        Id = subscription.Id ?? 0,
+                        Id = (int)(subscription.Id ?? 0),
                         State = subscription.State?.ToString(),
                         ProductName = subscription.Product?.Name,
                         ProductHandle = subscription.Product?.Handle,
